requirements.txt
asgiref==3.4.1
boto3==1.26.17
botocore==1.29.17
certifi==2022.6.15
charset-normalizer==2.0.9
click==8.0.4
colorgram.py==1.2.0
distlib==0.3.4
Django==4.0
django-bootstrap-modal-forms==2.2.0
django-cors-headers==3.13.0
django-filter==21.1
django-mathfilters==1.0.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.4.2
Flask==2.0.3
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
jmespath==1.0.1
MarkupSafe==2.1.1
Pillow==8.4.0
pipenv==2021.11.23
platformdirs==2.4.1
psycopg==3.0.10
psycopg2-binary==2.9.3
PyJWT==2.6.0
pyperclip==1.8.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.26.0
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.7
virtualenv==20.11.0
virtualenv-clone==0.5.7
Werkzeug==2.0.3
whitenoise==6.2.0
